requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.6
django-cors-headers>=4.1.0
django-filter>=23.2
drf-yasg>=1.21.5
pytest>=7.3.1
pytest-django>=4.5.2
black>=23.3.0
isort>=5.12.0
flake8>=6.0.0